# swagger.json
{
    "openapi": "3.0.1",
    "info": {
        "version": "1.0.0",
        "title": "Drust Daam APIs",
        "description": "User Project Application API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        }
    },
    "servers": [
        {
          "url": "https://drust-daam.herokuapp.com/api/",
          "description": "Live server"
        },
        {
          "url": "http://localhost:5000/api",
          "description": "Local server"
        }
      ],
...
    "schemes": [
        "http",
        "https"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "paths": {
...
        "/category": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Category" ],
                "summary": "Get all category in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Category" ],
                "summary": "Create a new category in system",
                "requestBody": {
                    "description": "Category Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
        "/category/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the category that we want to match",
                    "type": "string"
                }
            ],
            "delete": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Delete category with given ID",
                "tags": [ "Category" ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete category with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "404": {
                        "description": "Failed category not found."
                    }
                }
            },
            "put": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Update category with given ID",
                "tags": [ "Category" ],
                "requestBody": {
                    "description": "Category Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Category"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Category with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Category"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    },
                    "404": {
                        "description": "Failed unit not found."
                    }
                }
            }
        },
        "/area": {
            "get": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Area" ],
                "summary": "Get all area in the database.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Area"
                        }
                    }
                }
            },
             "post": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "Area" ],
                "summary": "Create a new area in system",
                "requestBody": {
                    "description": "Area Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Area"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
        "/area/{id}": {
            "parameters": [
                {
                    "name": "id",
                    "in": "path",
                    "required": true,
                    "description": "ID of the area that we want to match",
                    "type": "string"
                }
            ],
            "delete": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Delete area with given ID",
                "tags": [ "Area" ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Delete area with id",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Area"
                        }
                    },
                    "404": {
                        "description": "Failed area not found."
                    }
                }
            },
            "put": {
                "security": [{ "bearerAuth": [] }],
                "summary": "Update area with given ID",
                "tags": [ "Area" ],
                "requestBody": {
                    "description": "Area Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Area"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "description": "Area with new values of properties",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Area"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    },
                    "404": {
                        "description": "Failed unit not found."
                    }
                }
            }
        },
...
        "/signin": {
             "post": {
                "tags": [ "User" ],
                "summary": "Login in system",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/Login"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        },
        "/update/password": {
             "put": {
                "security": [{ "bearerAuth": [] }],
                "tags": [ "User" ],
                "summary": "Update password in system",
                "requestBody": {
                    "description": "User Object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/definitions/UpdatePassword"
                            }
                        }
                    }
                },
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/id"
                        }
                    },
                    "400": {
                        "description": "Failed Bad post data."
                    }
                }
            }
        }
    },
    "definitions": {
        "id": {
            "properties": {
                "uuid": {
                    "type": "string"
                }
            }
        },
...
        "Category": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                }
            }
        },
...
        "Area": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                }
            }
        },
...
        "Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "UpdatePassword": {
            "type": "object",
            "properties": {
                "currentPassword": {
                    "type": "string"
                },
                "updatePassword": {
                    "type": "string"
                }
            }
        }
    }
}